+++ other/brython/htdocs/brython.php
@@ -2,7 +2,7 @@
 $sourceCode = str_replace("</script>", "_BrythonWorkaroundForClosingScriptTag_", $_REQUEST['code']);
 
 // Si modifié, adapter '/themes/hugo-book/layouts/_default/baseof.html' de atlastk.org et zelbinium.q37.info.
-$version = "3.13.0";
+$version = "3.12.4";
 $eruda_version="3.4.0";
 
 if (!empty($_REQUEST['version'])) {