+++ tools/ucuq/proxy/core/messages.h
@@ -31,6 +31,7 @@ namespace messages {
     iUnknownProtocolVersion,
     iNoDeviceWithGivenTokenAndId,  // %1: token, %2: id
     iUnknownCaller,  // %1: caller
+    iUnknownScript, // %1: script name
     i_amount,
     i_Undefined
   };